split OppholdService
